+++ crates/resvg/examples/custom_usvg_tree.rs
@@ -49,10 +49,8 @@ fn main() {
     tree.calculate_abs_transforms();
     tree.calculate_bounding_boxes();
 
-    let rtree = resvg::Tree::from_usvg(&tree);
-
-    let pixmap_size = rtree.size.to_int_size();
+    let pixmap_size = tree.size.to_int_size();
     let mut pixmap = tiny_skia::Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();
-    rtree.render(tiny_skia::Transform::default(), &mut pixmap.as_mut());
+    resvg::render(&tree, tiny_skia::Transform::default(), &mut pixmap.as_mut());
     pixmap.save_png("out.png").unwrap();
 }