+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This changelog also contains important changes in dependencies.
 - `usvg::Node::abs_filters_bounding_box`.
 
 ### Changed
+- `resvg` renders `usvg::Tree` directly again. `resvg::Tree` is gone.
 - `usvg` no longer uses `rctree` for the nodes tree implementation.
   The tree is a regular `enum` now.
   - A caller no longer need to use the awkward `*node.borrow()`.
@@ -31,8 +32,10 @@ This changelog also contains important changes in dependencies.
 - Filters, clip paths, masks and patterns are stored as `Rc<RefCell<T>>` instead of `Rc<T>`.
   This is required for proper mutability since `Node` itself is no longer an `Rc`.
 - Rename `usvg::NodeKind` into `usvg::Node`.
+- Upgrade to Rust 2021 edition.
 
 ### Removed
+- `resvg::Tree`. No longer needed. `resvg` can render `usvg::Tree` directly once again.
 - `rctree::Node` methods. The `Node` API is completely different now.
 - `usvg::NodeExt`. No longer needed.
 - `usvg::Node::calculate_bbox`. Use `usvg::Node::abs_bounding_box`.